[Tests] added snapshot regression tests and modified test setup
