+++ pyproject.toml
@@ -29,6 +29,7 @@ dev = [
   "pytest",
   "pygame==2.5.0",
   "gymnasium[other]==1.2.0",
+  "syrupy==4.9.1",
 ]
 gh_ci = [
   "pygame==2.5.0",
